https://drive.google.com/file/d/1jPrlJy5mcRDvlyf5DDY-JdyVfDzjT6xw/view?usp=sharing
The real-time mask detection project is a system that uses computer vision and deep learning techniques to detect whether people are wearing masks or not directly in front of the camera. The program captures video using the computer’s camera, then detects faces and classifies them into three categories: properly wearing a mask, not wearing a mask, or wearing it incorrectly. A rectangle is displayed around each face with text indicating the category and the confidence level of the prediction. This project is useful for monitoring mask-wearing compliance in public places, hospitals, and offices, and also serves as an educational tool for applying computer vision and deep learning concepts to real-world data. The model was trained using the Face Mask Detection dataset from Kaggle and can be further developed to support faster models like YOLOv8 or graphical interfaces and mobile applications using TensorFlow Lite. Download the dataset from the link.